+++ src/Core/Entity/SyncParameter.php
@@ -74,9 +74,9 @@ class SyncParameter
         return new SyncParameter($name, SyncParameterType::UUID, $version, $isNullable);
     }
 
-    public static function createRelationOneToMany(string $name, array $relatedClass, int $version): self
+    public static function createRelationOneOfMany(array $relatedClass, int $version): self
     {
-        return new SyncParameter($name, SyncParameterType::RELATION_ONE_TO_MANY, $version, false, $relatedClass);
+        return new SyncParameter("relations_one_of_many", SyncParameterType::RELATION_ONE_TO_MANY, $version, false, $relatedClass);
     }
 
     // ------------------------------------------------------------------------