+++ tests/_Stubs/ParentFakeEntity.php
@@ -16,8 +16,6 @@ class ParentFakeEntity extends EntitySynchronizable
 
     const ATTR_ENUM = "value_enum";
 
-    const RELATED_CHILDREN = "children";
-
     const VERSION_NAME = 1;
 
     const VERSION_DEFAULT = 2;
@@ -33,7 +31,7 @@ class ParentFakeEntity extends EntitySynchronizable
             SyncParameter::createTimestamp(self::ATTR_COLOR, self::VERSION_DEFAULT),
             SyncParameter::createString(self::ATTR_VALUE_NULLABLE, self::VERSION_DEFAULT, true),
             SyncParameter::createEnum(self::ATTR_ENUM, self::ENUM_VALUES, self::VERSION_DEFAULT),
-            SyncParameter::createRelationOneToMany(self::RELATED_CHILDREN, [ChildFakeEntity::class], self::VERSION_CHILDREN)
+            SyncParameter::createRelationOneOfMany([ChildFakeEntity::class], self::VERSION_CHILDREN)
         ];
     }