+++ changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 2024-08-14
+- Se refactoriza el metodo de **SyncParameter::createRelationOneOfMany**.
 ## 2024-08-13   
 
 Se agrega soporte para el tipo de parametro de **_Enum_**.